requirements.txt
pyrogram
attrify
aiohttp==3.9.5
gpytranslate
rich==3.2.0
motor
openai==0.28
pytz
psutil
pykeyboard
python-dotenv
python-dateutil
py-tgcalls==2.0.2
pymongo
Pillow==9.5.0
pyromod==1.0
tgcrypto
telegraph
uvloop
wget
youtube-search-python
yt-dlp
google
search-engine-parser
requests
bs4
speedtest-cli